refactor(db): move tag list query to repository (DB-REPOSITORY-PILOT-NEXT-D1-A)
- 新增 tagsRepository.listByUser() 方法
- 在 types.ts 中添加 TagWithCount 类型
- 修改 routes/tags.ts 调用 Repository
- 保持 SQL 语义和返回结构完全不变
